Split API implementations into apps.
